delete scrap
