+++ deploy/index.ts
@@ -1,4 +1,4 @@
-import { multiswap } from "./multiswapTesting";
+import { multiswap } from "./multiswap";
 
 
 const main = async () => {
